refactored: deleted unused files
[] changed directory for better management
[] deleted unused files
[] temporarily removed redirection after logging in
