--- app/src/main/java/com/example/mad_finals_wurmple/menuActivity.kt
@@ -1,4 +1,0 @@
-package com.example.mad_finals_wurmple
-
-class menuActivity {
-}
